move resolution-api to handlers
